+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### WIP
 
 - Beautified system emails
+- Added admin theme font subset latin, latin-ext
 
 ### 0.8.4 (July 25, 2014)